@@ -1,30 +1,28 @@
 buildscript {
     repositories {
-        maven { url 'https://maven.fabric.io/public' }
+        google()
+        mavenCentral()
+    }
+    dependencies {
+        classpath 'com.google.gms:google-services:4.3.5'
+        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.5.2'
     }
 
-    dependencies {
-        classpath 'io.fabric.tools:gradle:1.+'
-    }
 }
 
 apply plugin: 'com.android.application'
-
-repositories {
-    maven { url 'https://maven.fabric.io/public' }
-}
-
-apply plugin: 'io.fabric'
+apply plugin: 'com.google.gms.google-services'
+apply plugin: 'com.google.firebase.crashlytics'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 30
 
     defaultConfig {
         applicationId "com.spicymango.fanfictionreader"
-        minSdkVersion 14
-        targetSdkVersion 28
-        versionCode 93
-        versionName "1.58a"
+        minSdkVersion 16
+        targetSdkVersion 30
+        versionCode 99
+        versionName "1.63"
     }
     lintOptions {
         abortOnError true
@@ -49,7 +47,9 @@ android {
             // 2. buildConfigField is used by static constant in StoryProvider.java
             // Note: owner refers to current script block
             addConstantTo(owner, "provider_authority", "com.spicymango.fanfictionreader.debug.provider")
+            minifyEnabled true
             ext.enableCrashlytics = false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-project-debug.pro'
         }
         release {
             addConstantTo(owner, "provider_authority", "com.spicymango.fanfictionreader.provider")
@@ -67,16 +67,16 @@ android {
 
 dependencies {
     implementation project(':kolavar')
-    implementation 'io.reactivex.rxjava2:rxandroid:2.1.0'
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.2'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
+    implementation 'io.reactivex.rxjava3:rxandroid:3.0.0'
+    implementation 'io.reactivex.rxjava3:rxjava:3.0.11'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation "androidx.fragment:fragment:1.3.3"
+    implementation 'com.google.android.material:material:1.3.0'
     implementation 'org.jsoup:jsoup:1.11.3'
     implementation 'org.apache.commons:commons-collections4:4.2'
     implementation 'org.apache.commons:commons-lang3:3.8.1'
-    implementation ('com.crashlytics.sdk.android:crashlytics:2.9.8@aar') {
-        transitive = true
-    }
+    implementation 'com.google.firebase:firebase-crashlytics:17.4.1'
+    implementation 'com.google.firebase:firebase-analytics:18.0.3'
 }
 
 // If the signing configurations are defined in the local gradle.properties, add them to the script
@@ -85,4 +85,6 @@ if (project.hasProperty('key.alias')) {
     android.signingConfigs.release.keyPassword = project.property('key.alias.password')
     android.signingConfigs.release.storeFile = file(project.property('key.store'))
     android.signingConfigs.release.storePassword = project.property('key.store.password')
+} else {
+    android.buildTypes.release.signingConfig = null
 }
