@@ -13,12 +13,12 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.content.Loader;
-import android.support.v4.view.MenuItemCompat;
-import android.support.v7.widget.SearchView;
-import android.support.v7.widget.SearchView.OnQueryTextListener;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.FragmentActivity;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.Loader;
+import androidx.appcompat.widget.SearchView;
+import androidx.appcompat.widget.SearchView.OnQueryTextListener;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -46,6 +46,7 @@ public class SearchAuthorActivity extends BaseActivity<MenuObject> implements On
 	private SearchLoader<MenuObject> mLoader; 
 	private SearchView sView;
 	
+	@NonNull
 	@Override
 	public Loader<List<MenuObject>> onCreateLoader(int id, Bundle args) {
 		mLoader = new AuthorLoader(this, args);
@@ -59,7 +60,7 @@ public class SearchAuthorActivity extends BaseActivity<MenuObject> implements On
 		MenuInflater inflater = getMenuInflater();
 		inflater.inflate(R.menu.search_menu, menu);
 
-		sView = (SearchView) MenuItemCompat.getActionView(menu.findItem(R.id.search));
+		sView = (SearchView) menu.findItem(R.id.search).getActionView();
 		sView.setOnQueryTextListener(this);
 		return super.onCreateOptionsMenu(menu);
 	}
@@ -80,8 +81,8 @@ public class SearchAuthorActivity extends BaseActivity<MenuObject> implements On
 	}
 
 	@Override
-	public void onLoadFinished(Loader<List<MenuObject>> loader,
-			List<MenuObject> data) {
+	public void onLoadFinished(@NonNull Loader<List<MenuObject>> loader,
+							   List<MenuObject> data) {
 		super.onLoadFinished(loader, data);
 		mLoader = (SearchLoader<MenuObject>) loader;
 		supportInvalidateOptionsMenu();
@@ -89,11 +90,10 @@ public class SearchAuthorActivity extends BaseActivity<MenuObject> implements On
 	
 	@Override
 	public boolean onOptionsItemSelected(MenuItem item) {
-		switch (item.getItemId()) {
-		case R.id.filter:
+		if (item.getItemId() == R.id.filter){
 			((Filterable) mLoader).onFilterClick(this);
 			return true;
-		default:
+		}{
 			return super.onOptionsItemSelected(item);
 		}
 	}
@@ -131,7 +131,7 @@ public class SearchAuthorActivity extends BaseActivity<MenuObject> implements On
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		getSupportLoaderManager().initLoader(0, savedInstanceState, this);
+		LoaderManager.getInstance(this).initLoader(0, savedInstanceState, this);
 	}
 	
 	private static final class AuthorLoader extends SearchLoader<MenuObject> implements Filterable{
