@@ -1,14 +1,13 @@
 package com.spicymango.fanfictionreader.dialogs.backup;
 
 import com.spicymango.fanfictionreader.R;
-import com.spicymango.fanfictionreader.dialogs.backup.RestoreDialog;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.DialogInterface;
 import android.content.DialogInterface.OnClickListener;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 
 
 public class RestoreDialogConfirmation extends DialogFragment implements OnClickListener {
