@@ -3,9 +3,9 @@ package com.spicymango.fanfictionreader.menu.authormenu;
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentActivity;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.FragmentActivity;
 
 import com.spicymango.fanfictionreader.R;
 import com.spicymango.fanfictionreader.menu.BaseLoader;
