@@ -5,7 +5,7 @@ import java.util.List;
 import com.spicymango.fanfictionreader.R;
 
 import android.content.Context;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
