@@ -1,7 +1,7 @@
 package com.spicymango.fanfictionreader.menu.mainmenu;
 
 import android.content.Context;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
