@@ -2,7 +2,7 @@ package com.spicymango.fanfictionreader.menu.reviewmenu;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.Date;
 
