@@ -25,7 +25,7 @@ import com.spicymango.fanfictionreader.util.Story;
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 
 final class StoryMenuLoaders {
 
