@@ -1,6 +1,6 @@
 package com.spicymango.fanfictionreader.util;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import org.jsoup.Connection;
 import org.jsoup.Jsoup;
