@@ -1,18 +1,26 @@
 package com.spicymango.fanfictionreader.services;
 
+import android.annotation.SuppressLint;
 import android.content.ContentResolver;
 import android.content.Context;
 import android.content.UriMatcher;
 import android.database.Cursor;
 import android.net.Uri;
+import android.os.Handler;
 import android.text.TextUtils;
+import android.util.Log;
 import android.util.SparseArray;
+import android.webkit.CookieManager;
+import android.webkit.WebResourceError;
+import android.webkit.WebResourceRequest;
+import android.webkit.WebResourceResponse;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
 
 import com.spicymango.fanfictionreader.activity.Site;
 import com.spicymango.fanfictionreader.provider.SqlConstants;
 import com.spicymango.fanfictionreader.provider.StoryProvider;
 import com.spicymango.fanfictionreader.util.FileHandler;
-import com.spicymango.fanfictionreader.util.JsoupUtil;
 import com.spicymango.fanfictionreader.util.Parser;
 import com.spicymango.fanfictionreader.util.Sites;
 import com.spicymango.fanfictionreader.util.Story;
@@ -23,6 +31,11 @@ import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
 import java.io.IOException;
+import java.net.CookieHandler;
+import java.net.CookieStore;
+import java.net.HttpCookie;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.text.ParseException;
 import java.util.Date;
 import java.util.regex.Matcher;
@@ -70,13 +83,11 @@ class DownloaderFactory {
 	 * @param context The current context
 	 * @return The downloader that works for the requested URI
 	 */
-	static Downloader getInstance(Uri uri, Context context) {
-		switch (URI_MATCHER.match(uri)) {
-			case FAN_FICTION:
-				return new FanFictionDownloader(uri, context);
-			default:
-				throw new UnsupportedOperationException("Downloader Factory does not support the Uri: " + uri);
+	static Downloader getInstance(Uri uri, Context context, WebView webView) {
+		if (URI_MATCHER.match(uri) == FAN_FICTION) {
+			return new FanFictionDownloader(uri, context, webView);
 		}
+		throw new UnsupportedOperationException("Downloader Factory does not support the Uri: " + uri);
 	}
 
 	/**
@@ -136,16 +147,18 @@ class DownloaderFactory {
 		 *
 		 * @param lastPageRead The last chapter read by the user
 		 * @param scrollOffset The user's position along the chapter
+		 * @param saveChapters True if both chapters and metadata should be saved, false if only
+		 *                           metadata should be updated.
 		 * @throws IOException If writing to the device memory fails
 		 */
-		void saveStory(int lastPageRead, int scrollOffset) throws IOException;
+		void saveStory(int lastPageRead, int scrollOffset, boolean saveChapters) throws IOException;
 
 		/**
 		 * Instructs the downloader to skip pre-existing chapters.
 		 */
 		void EnableIncrementalUpdating();
 
-		void downloadIfMissing() throws IOException, ParseException, StoryNotFoundException;;
+		void downloadIfMissing() throws IOException, ParseException, StoryNotFoundException;
 	}
 
 	private static final class FanFictionDownloader implements Downloader{
@@ -191,9 +204,21 @@ class DownloaderFactory {
 		 */
 		private final Context mContext;
 
-		private FanFictionDownloader(Uri uri, Context context) {
+		/**
+		 *  A web view used to bypass Cloudflare's captcha
+		 *  */
+		private final WebView mWebView;
+
+		private final Object mutex;
+		private String mHtmlFromWebView;
+
+
+		@SuppressLint("AddJavascriptInterface")
+		private FanFictionDownloader(Uri uri, Context context, WebView webView) {
 			mContext = context;
 			mCurrentPage = 1;
+			mutex = new Object();
+			mHtmlFromWebView = null;
 
 			mText = new SparseArray<>();
 
@@ -202,6 +227,29 @@ class DownloaderFactory {
 			Matcher idMatcher = PATTERN_FF.matcher(uri.toString());
 			if (!idMatcher.find()) throw new IllegalArgumentException("The uri " + uri + " is not valid");
 			mStoryId = Long.parseLong(idMatcher.group(1));
+
+			// Initialize the webView's callbacks
+			mWebView = webView;
+
+			// Try to load the document. This must be done from the main thread
+			final Handler mainHandler = new Handler(mContext.getMainLooper());
+			final Runnable runnable = () -> {
+				mWebView.setWebViewClient(new CustomWebView());
+				mWebView.addJavascriptInterface(new JavascriptListener(), "HTMLOUT");
+				synchronized (mutex){
+					mutex.notify();
+				}
+			};
+
+			// Wait for the WebView to finish loading
+			synchronized (mutex){
+				try {
+					mainHandler.post(runnable);
+					mutex.wait();
+				} catch (InterruptedException e){
+					Log.e(this.getClass().getSimpleName(), "Thread interrupted", e);
+				}
+			}
 		}
 
 		@Override
@@ -269,7 +317,29 @@ class DownloaderFactory {
 		public void downloadChapter() throws IOException, ParseException, StoryNotFoundException {
 			final String url = "https://www.fanfiction.net/s/" + mStoryId + "/"
 					+ mCurrentPage + "/";
-			final Document document = JsoupUtil.safeGet(url, "Mozilla/5.0");
+
+			// Try to load the document. This must be done from the main thread
+			final Handler mainHandler = new Handler(mContext.getMainLooper());
+			final Runnable runnable = () -> mWebView.loadUrl(url);
+			mHtmlFromWebView = null;
+
+			// Wait for the WebView to finish loading
+			synchronized (mutex){
+				mainHandler.post(runnable);
+				try {
+					while (mHtmlFromWebView == null){
+						mutex.wait();
+					}
+				} catch (InterruptedException e){
+					throw new IOException();
+				}
+			}
+
+			if (mHtmlFromWebView.equalsIgnoreCase("404")){
+				throw new IOException();
+			}
+
+			final Document document = Jsoup.parse(mHtmlFromWebView, url);
 
 			// On the first run, update the mStory variable
 			if (mCurrentPage == 1) {
@@ -310,11 +380,13 @@ class DownloaderFactory {
 		}
 
 		@Override
-		public void saveStory(int lastPageRead, int scrollOffset) throws IOException{
+		public void saveStory(int lastPageRead, int scrollOffset, boolean saveChapters) throws IOException{
 			// Write each of the chapters downloaded into the file system
-			for (int i = 0; i < mText.size(); i++) {
-				final int key = mText.keyAt(i);
-				FileHandler.writeFile(mContext, mStoryId, key, mText.get(key));
+			if (saveChapters){
+				for (int i = 0; i < mText.size(); i++) {
+					final int key = mText.keyAt(i);
+					FileHandler.writeFile(mContext, mStoryId, key, mText.get(key));
+				}
 			}
 
 			// By default, the Story object has a date added equal to 0 upon creation. If this value
@@ -397,7 +469,7 @@ class DownloaderFactory {
 
 			Matcher matcher = PATTERN_FF.matcher(authorElement.attr("href"));
 			if (!matcher.find()) return null;
-			int authorId = Integer.valueOf(matcher.group(1));
+			int authorId = Integer.parseInt(matcher.group(1));
 
 			Element summaryElement = document.select("div#profile_top > div").first();
 			if (summaryElement == null) return null;
@@ -442,6 +514,93 @@ class DownloaderFactory {
 			builder.setCompleted(attributes.text().contains("Complete"));
 
 			return builder.build();
+		}
+
+
+		private class JavascriptListener{
+			@android.webkit.JavascriptInterface
+			public void processHTML(String html) {
+				mHtmlFromWebView = html;
+				synchronized (mutex) {
+					mutex.notify();
+				}
+			}
+		}
+
+		private class CustomWebView extends WebViewClient {
+			private final Object internalMutex = new Object();
+			private boolean mHasLoaded;
+
+			@Override
+			public void onReceivedError(WebView view, WebResourceRequest request,
+										WebResourceError error) {
+				super.onReceivedError(view, request, error);
+
+				mHtmlFromWebView = "404";
+				synchronized (mutex) {
+					mutex.notify();
+				}
+			}
+
+			@Override
+			public void onReceivedHttpError(WebView view, WebResourceRequest request,
+											WebResourceResponse errorResponse) {
+				super.onReceivedHttpError(view, request, errorResponse);
+
+				// If an HttpError is received, it may be a "Wait a few seconds" Cloudflare page.
+				// Use the internalMutex to wait 5 seconds before assuming an error.
+				mHasLoaded = false;
+				new Thread(()->{
+					synchronized (internalMutex){
+						try {
+							internalMutex.wait(5*1000);
+						} catch (InterruptedException ignored) {}
+					}
+
+					if (!mHasLoaded){
+
+						// Even after waiting, the error persists. Return a connection error.
+						mHtmlFromWebView = "404";
+						synchronized (mutex) {
+							mutex.notify();
+						}
+					}
+				}).start();
+			}
+
+			@Override
+			public void onPageFinished(WebView view, String url) {
+				// Notify mutex that the Cloudflare "Wait a few seconds" page has loaded.
+				mHasLoaded = true;
+				synchronized (internalMutex){
+					internalMutex.notify();
+				}
+
+				// Pass the cookies from the webView to the cookie storage
+				final CookieManager manager = CookieManager.getInstance();
+				final String httpCookieHeader = manager.getCookie(url);
+
+				if (httpCookieHeader != null){
+					try {
+						final URI uri = new URI("https://fanfiction.net/");
+						final CookieStore cookieStore = ((java.net.CookieManager) CookieHandler.getDefault()).getCookieStore();
+
+						for (String cookieString : httpCookieHeader.split(";")){
+							final String[] splitCookie = cookieString.split("=");
+							HttpCookie cookie = new HttpCookie(splitCookie[0], splitCookie[1]);
+							cookieStore.add(uri, cookie);
+						}
+
+					} catch (URISyntaxException e) {
+						Log.d(getClass().getSimpleName(), "Failed to parse URI =" + url, e);
+					}
+				}
+
+				// Retrieve the html code.
+				view.loadUrl("javascript:window.HTMLOUT.processHTML('<html>'+document.getElementsByTagName('html')[0].innerHTML+'</html>');");
+
+				super.onPageFinished(view, url);
+			}
 		}
 
 	}
